Big update. New update format, Update EA-15W49A
I have a new way I will be labeling difference versions of this mod.
This update is the start of the new layout.  I am adapting the Minecraft
snapshot way of updating, as I make several changes and updates a week.
This could change in the future.

I went through the rest of the files and did much needed code tweaking
and clean up.

I completely implemented Ore WorldGen for Arcanite Ore.

I finished up all of the Arcanite Blocks and Innate Arcanite Blocks
metas, complete with textures.

I restructured the Arcanite Fence, preping it for a meta conversion as
well. That will not be until much later. I will most likely get walls,
slabs, and stairs working first before this.

Started to add placeholders for a new multi-ore called Ebon Ore.  Ebon
Ore has a chance of dropping multiple randomized ores and gems

Started to add placeholders for a multipurpose tool called The
Ebonheart.  This tool can mine, cut wood, and be used as a shovel.  It
will have the power of Diamond but with more durability.

A lot of misc things done.