EA-15W50C
Huge update again.  A lot of code cleanup.

My first test tile entity is in the game.  This is a 9 slot box that
saves what is up into it on game restart.  This is the foundation to
other tile entities.

The main Gui structure is finished.  I am able to make Gui's for items.
That being said...

The Item Alchemic Compendium is added and has a skin!  When held in you
hand and right-click is pressed, a basic gui screen comes up.  this is
the foundation of the Compendium.

The Configuration structure has been built.  The config is currently
only for a basic crafting recipe change, but I need something simple to
test it.  you can change the boolean in the ea.cfg file that is created.
This took a lot of work to have the config file used, and to save the
changes.  A Gui for in-game changes through the mod list options are
working as well, however it does require a restart to take affect.  This
is by design.

Reworked and re-enabled the Crafting recipes file.  It now gives me more
options to enable or disable shaped, shapeless or smelting recipes
separately.

Started to do a little more work with the multiblock structure, the
mystic furnace- WIP.

Added mcmod.info and created a basic logo.