EA-15W50F
Changed the spelling of "katchin" to the phoenetic spelling of
"katcheen".  This was actually huge task as the ore is referenced
multiple times all through out the mod.  This will be better in the
future when people say it correctly.

Added blocks: Draconium Block, Velious Block, Katcheen Block, Necrocite
Block, and Soularite Block.

Re-made various items, ores, and blocks textures.