15W51B
Added item Shimmering Disc which when combined to a particular dye,
creates a music disc.

Created a new config option enable or disable the crafting of music
discs!

Adjusted and organized a lot of code. Made classes for versatile and
redefined things.

Made new textures for Draconium, Velious, and Arcanite.