Pre v1.0.9a
- New sets of blocks have been added, variations of Glowstone and
Obsidian.

- Changed Block and Item Creative Tabs icons.

- Slabs did not work at all.  You could not get any items back when
harvesting them and the mmb didn't work.  Everything works 100% now.

- Ebonheart Gemstones now stack up to a Max of 16.

- Completely reworked all blocks hardness values. I made all crafted
blocks easy to harvest, since it is less annoying to build with them.
All ores are varing in hardness starting from the weakest ore Draconium,
and ending with the strongest ore Ebon Ore.  Draconium-Arcanite can be
harvested with iron or better, and Katcheen + can be harvested with
diamond or better.

-All blocks have new resistance values.