v1.0.13
Created ladders for all block types.

Torches are completely done.

created a block of leather.

Changed blocks of leather and paper to be orientable. Because of this,
you can rearrange the block in different directions.

Messed with doors, still not working.

Crafting recipes for all items are in the game.