v1.1.32
Updated a lot of stuff.  More revision bumps.

New models have been created for Pillars, Bricks, Ladders, and Torches.
Let me know how you like the new models!

Ebonheart now is properly named "Block of Ebonheart" and has a crafting
recipe.  This is going to be used as a storage block for "Ebonheart".

Adjusted a lot of block settings for all basic blocks to improve
transparency, shading, and structure of all new models.

New textures have been implemented for a lot of blocks to help with the
new models.

A lot of code cleanup.  Disabled "Tier Catalysts" as they server no
purpose.