v1.3.1
- Changed how most mod blocks reacted with ladders and torches.  Most
basic blocks besides the Brick types can not pass redstone current, have
torches, and ladders on them.

Note: When I tried to make this change on Bricks, there were dark
shadows in the seems of the bricks when next to other blocks.  I could
set this feature but have that graphical glitch.  I will try to resolve
this.

- Shimmering disc will now play an 8 bit version of "Lost Woods" from
Zelda - Ocarina of Time (edited in audacity).