v1.3.2
The Item Ebonheart has changed.  The tooltip was updated.  When holding
it in your hand, you must now hold shift + right click to throw the
item.  The Item now renders correctly.

Made a new system for rendering entities.

With bricks, I decided to go with utility over graphics.  That means the
weird black shadows will appear when placed next to whole blocks.  It is
very minor and is easily overlooked.  There is nothing that can be done
because it is all on the minecraft rendering engine.

Correctly separated rendering from registration.

A lot more code cleanup.