1.3.2c
Arcanite armor is now repairable with arcanite in an anvil.

Katcheen armor is now repairable with katcheen in an anvil.

New Item Dragon Fruit is now in the game.  This fruit is only found in
random chests, and the creative Items Tab.  It heals 6 hearts and has a
saturation a little under A golden apple.