v1.4.0
Updated my meta ModelBakery depreciated code.

Added a New Ore, Ebonheart Ore.  It spawns only in the nether.

Changed the item Ebonheart to be a regular item like the other
resources.

Created a new item, Enchanted Ebonheart.  This item has a new skin. This
is not the throwable kind -WIP, still need to make a crafting recipe for
it.

Innate Mystic Block was renamed to Mystic Cobblestone.

Both Mystic Cobblestone and Mystic Stone are now in the mod 100%.  They
are craftable.

Reorganized creating tabs Block and Item locations.

Reorganized code to make it easier to look at.

Recipes added, Mystic Cobblestone, Mystic Stone, made the ability to dye
all Mystic Cobblestone/Stone blocks.  Make it so all said dyed blocks
can be reconverted back, changed all of the Ultimate weapon recipes to
use an Iron ingot instead of Ebonheart for the handle.

Remember, the new achievements for the new music discs.