v2.2.9
Made adjustments to configs.  Please delete your old config to minimize
confusion.

Configs are colored in the in-game GUI!  Have fun!

The two new Items, Necrosis and Consecration are fully working.  Ready
to go.

Designed a permanent way to handle armor buffs.  I created an event
handler to monitor the buffs and once a piece is missing, will correctly
remove the buffs.  There was a bug that could allow players to keep
their buffs until another armor piece was equip.  With this change, I
set it so that buffs will work in creative again.

Added multi-color flame animations to Enchanted Ebonheart upon impact.

Adjusted Celestial Armor Rune animations to give off a hue of random,
Soularite colored runes.  Increased the size of the runes as well.

Redesigned the Celestial armor to resemble the material it was created
out of, Soularite.

Adjusted a lot of various textures.

Rearranged a few items in the creative tabs to make the mod items more
visually pleasing.  Can be seen in new maps.

Code cleanup.  A lot of it again.